requirements.txt
Pillow==10.3.0
streamlit==1.33.0
transformers==4.40.0
